Update: Allow visitors to toggle dark theme
Add a new sticky button to the bottom-right corner allowing
visitors to toggle dark mode on their end. For first-time
visitors, the value defaults to the value set in
`_configs.yml/darkmode`.

Stores a cookie for the theme that will be used in subsequent
visits. In case multiple sessions to the website are opened,
changing the mode in any once instace will reflect across all
tabs.

This feature can be turned off by modifying the value of
`_config.yml/darktoggle`.

Move the existing contents of `/assets/index.js` as an inline
script.